Ajustes

+++ app/Controllers/Http/BooksController.ts
@@ -16,6 +16,11 @@ export default class BooksController {
     public async update ({ request, response }: HttpContextContract){
         const book = await Book.findByOrFail('id', request.param('id'))
         const bookPayload = request.only(['title', 'author', 'cover', 'category', 'quantity'])
+
+        if(!book){
+            return response.notFound({message: 'Book not found'})
+        }
+
         book.merge(bookPayload)
         await book.save()
         return response.ok({book})
@@ -23,16 +28,26 @@ export default class BooksController {
 
     public async destroy ({ request, response }: HttpContextContract){
         const book = await Book.findByOrFail('id', request.param('id'))
+
+        if(!book){
+            return response.notFound({message: 'Book not found'})
+        }
+
         await book.delete()
         return response.ok({book})
     }
 
-    public async show ({ response, request }: HttpContextContract){
+    public async findOne ({ response, request }: HttpContextContract){
         const book = await Book.findByOrFail('id', request.param('id'))
+
+        if(!book){
+            return response.notFound({message: 'Book not found'})
+        }
+
         return response.ok({book})
     }
     
-    public async showAll ({ response }: HttpContextContract){
+    public async findAll ({ response }: HttpContextContract){
         const books = await Book.all()
         return response.ok({books})
     }
